Added local storage
Store is a simple class which can be initialized with a name and initial
value. It has two methods: dump, which stringifies the data, and dump
which converts it back to a data structure.

Each time we run a mutating function on our Riot component, we update
the store. When the page is reloaded, we initialize the store and if it
is defined use it for the initial issues.